Publish the delta algorithm seam and address inputs in 64 bits
DeltaAlgorithm, DeltaScanner and DeltaSink are public, and the patcher
constructors take one, so a third party can replace how bytes are matched
without touching the patch format. Nothing about the search reaches the
container - the COPY/DIFF/ADD/RUN set is fixed - so a patch built by a
custom algorithm is restored by the stock patcher.

Inputs are now a SeekableSource rather than a ByteArray. A delta is not a
sequential transform: a copy names an arbitrary source offset and applying
one reads it back, so the right abstraction is random access, not a stream.
fileSource(path) opens a file for it on every target with a file system -
RandomAccessFile on the JVM, stdio on native, fs.readSync on Node - and the
browser, which has none, says so. Checksums now run straight off a source a
chunk at a time, so a file is no longer read whole to be verified.

Sizes and delta offsets become 64-bit, format v2. For values inside Int
range the varint bytes are exactly what v1 wrote, so v1 patches still
apply; only the declared ceiling moved.

The bundled RollingHashAlgorithm still indexes a ByteArray, so a patcher
materializes both inputs once before encoding and refuses anything above
2 GB with a clear error. A bounded-memory search is the next step and needs
no format change when it lands.

+++ library/src/commonMain/kotlin/com/linroid/kiff/PatchInfo.kt
@@ -3,13 +3,13 @@ package com.linroid.kiff
 data class PatchInfo(
   val patcher: PatcherId,
   val formatVersion: Int,
-  val sourceSize: Int,
+  val sourceSize: Long,
   val sourceCrc32: UInt,
-  val targetSize: Int,
+  val targetSize: Long,
   val targetCrc32: UInt,
-  val patchSize: Int
+  val patchSize: Long
 ) {
   /** Patch size as a fraction of the target size; smaller is better. */
   val ratio: Double
-    get() = if (targetSize == 0) 0.0 else patchSize.toDouble() / targetSize.toDouble()
+    get() = if (targetSize == 0L) 0.0 else patchSize.toDouble() / targetSize.toDouble()
 }
